Oops....

+++ src/asset_handler.cpp
@@ -115,7 +115,7 @@ int main(int arg_len, char **args)
 					header.data_size = header.image.width * header.image.height * header.image.color_depth;
 					printf("%s (image), %dx%dx%d, size: %d\n", file_path, 
 							header.image.width, header.image.height, 
-							header.image.color_depth, (uint32) header.data_size);
+							header.image.color_depth, header.data_size);
 					break;
 				}
 			case (BAT_SOUND):
